migrate mocks to test utils
