+++ plugins/cost-insights/src/components/ProjectGrowthAlertCard/ProjectGrowthAlertCard.test.tsx
@@ -17,8 +17,8 @@
 import React from 'react';
 import { renderInTestApp } from '@backstage/test-utils';
 import { ProjectGrowthAlertCard } from './ProjectGrowthAlertCard';
-import { createMockProjectGrowthData } from '../../utils/mockData';
 import {
+  createMockProjectGrowthData,
   MockCurrencyProvider,
   MockConfigProvider,
   MockBillingDateProvider,
