+++ plugins/cost-insights/src/testUtils/filters.ts
@@ -14,16 +14,9 @@
  * limitations under the License.
  */
 
-export {
-  costInsightsPlugin,
-  costInsightsPlugin as plugin,
-  CostInsightsPage,
-  CostInsightsProjectGrowthInstructionsPage,
-  CostInsightsLabelDataflowInstructionsPage,
-} from './plugin';
-export { ExampleCostInsightsClient } from './example';
-export { BarChart, LegendItem, CostGrowth } from './components';
-export { MockConfigProvider, MockCurrencyProvider } from './testUtils';
-export * from './api';
-export * from './alerts';
-export * from './types';
+import { Duration, ProductFilters } from '../types';
+import { MockProductTypes } from './products';
+
+export const MockProductFilters: ProductFilters = Object.keys(
+  MockProductTypes,
+).map(productType => ({ duration: Duration.P30D, productType }));
