+++ plugins/cost-insights/src/testUtils/index.ts
@@ -13,5 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+export * from './alerts';
+export * from './config';
+export * from './filters';
+export * from './loading';
+export * from './mockData';
+export * from './products';
 export * from './providers';
+export * from './testUtils';
